# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "0.0.9"
  title: nodeTestApp16
host: dockerHostIp:10011
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /{locale}/graph/relationships/create:
    x-swagger-router-controller: context_html_web_graph_relationships
    get:
      tags:
        - web
        - graph
        - relationships
      summary: Create relationship page (GET)
      description: Returns html page for a relationship creation form
      produces: 
        - text/html
      operationId: getCreate
      parameters:
        - $ref: "#/parameters/locale"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      tags:
        - web
        - graph
        - relationships
      summary: Relationship create target (POST)
      description: Creates a new relationship entry via Web (HTML in Browser) and then redirects to read this relationship (Code '302' is the default for a successful redirect in express.js)
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: create
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/source"
        - $ref: "#/parameters/type"
        - $ref: "#/parameters/target"
        - $ref: "#/parameters/property"
      responses:
        "302":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
  /{locale}/graph/relationships/update/{id}:
    x-swagger-router-controller: context_html_web_graph_relationships
    get:
      tags:
        - web
        - graph
        - relationships
      summary: Update relationship page (GET)
      description: Returns html page for a relationship update form
      produces: 
        - text/html
      operationId: getUpdate
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      tags:
        - web
        - graph
        - relationships
      summary: Relationship update target (POST)
      description: Updates a relationship entry via Web (HTML in Browser) without redirect and only for ACTED_IN and REVIEWED relationships
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: update
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
        - $ref: "#/parameters/type"
        - $ref: "#/parameters/property"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
  /api/movies:
    x-swagger-router-controller: context_json_api_movies
    post:
      tags:
        - api
        - nodes
        - movies
      summary: Creates movie
      description: Creates a new movie entry via JSON API
      operationId: create
      parameters:
        - name: movie
          required: true
          in: body
          description: a new movies details
          schema:
            $ref: "#/definitions/MovieBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MovieBody"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - api
        - nodes
        - movies
      summary: Reads several movies
      description: Returns a list of movies via JSON API with optional filter by released year passed in as inQeryParam
      operationId: readBulk
      parameters:
        - name: released
          in: query
          description: filters movies by released date
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MoviesListBody"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /{locale}/movies/create:
    x-swagger-router-controller: context_html_web_movies
    get:
      tags:
        - web
        - nodes
        - movies
      summary: Create movie page (GET)
      description: Returns a page with movie creation from
      produces: 
        - text/html
      operationId: getCreate
      parameters:
        - $ref: "#/parameters/locale"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      tags:
        - web
        - nodes
        - movies
      summary: Movie create target (POST)
      description: Creates a new movie entry via Web (HTML in Browser) and then redirects to read this movie (Code '302' is the default for a successful redirect in express.js)
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: create
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/title"
        - $ref: "#/parameters/tagline"
        - $ref: "#/parameters/released"
      responses:
        "302":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
  /{locale}/movies:  
    x-swagger-router-controller: context_html_web_movies
    get:
      tags:
        - web
        - nodes
        - movies
      summary: Reads several movies
      description: Returns a page with a list of movies with optional filter by released year passed in as inQeryParam
      produces: 
        - text/html
      operationId: readBulk
      parameters:
        - $ref: "#/parameters/locale"
        - name: released
          in: query
          description: Filters movies by released date
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
  /{locale}/movies/update/{id}:
    x-swagger-router-controller: context_html_web_movies
    get:
      tags:
        - web
        - nodes
        - movies
      summary: Update movie page (GET)
      description: Returns html page with a movie update form by a given id via inPathParam
      produces: 
        - text/html
      operationId: getUpdate
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      summary: Movie update target (POST)
      description: Updates a single movie entry by a given id via inPathParam (without redirect)
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: update
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
        - $ref: "#/parameters/title"
        - $ref: "#/parameters/tagline"
        - $ref: "#/parameters/released"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
  /api/persons:
    x-swagger-router-controller: context_json_api_persons
    post:
      tags:
        - api
        - nodes
        - persons
      summary: Creates person
      description: Creates a new person entry via JSON API
      operationId: create
      parameters:
        - name: person
          required: true
          in: body
          description: a new persons details
          schema:
            $ref: "#/definitions/PersonBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PersonBody"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - api
        - nodes
        - persons
      summary: Reads several persons
      description: Returns a list of persons via JSON API with optional filter by birth year passed in as inQeryParam
      operationId: readBulk
      parameters:
        - name: born
          in: query
          description: filters persons by birth date
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PersonsListBody"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /{locale}/persons/create:
    x-swagger-router-controller: context_html_web_persons
    get:
      tags:
        - web
        - nodes
        - persons
      summary: Create person page (GET)
      description: Returns html page with person creation form
      produces: 
        - text/html
      operationId: getCreate
      parameters:
        - $ref: "#/parameters/locale"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      tags:
        - web
        - nodes
        - persons
      summary: Person create target (POST)
      description: Creates a new person entry via Web (HTML in Browser) and then redirects to read this movie (Code '302' is the default for a successful redirect in express.js)
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: create
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/name"
        - $ref: "#/parameters/born"
      responses:
        "302":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
  /{locale}/persons:
    x-swagger-router-controller: context_html_web_persons
    get:
      tags:
        - web
        - nodes
        - persons
      summary: Reads several persons
      description: Returns a list of persons with optional filter by birth year passed in as inQeryParam
      produces: 
        - text/html
      operationId: readBulk
      parameters:
        - $ref: "#/parameters/locale"
        - name: born
          in: query
          description: filters persons by birth date
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
  /{locale}/persons/update/{id}:
    x-swagger-router-controller: context_html_web_persons
    get:
      tags:
        - web
        - nodes
        - persons
      summary: Update person page (GET)
      description: Returns html page with a person update form
      produces: 
        - text/html
      operationId: getUpdate
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
    post:
      tags:
        - web
        - nodes
        - persons
      summary: Person update target (POST)
      description: Updates a single person entry by a given id via inPathParam (without redirect)
      consumes:
        - application/x-www-form-urlencoded
      produces: 
        - text/html
      operationId: update
      parameters:
        - $ref: "#/parameters/locale"
        - $ref: "#/parameters/id"
        - $ref: "#/parameters/name"
        - $ref: "#/parameters/born"
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponseWeb"
...
parameters:
...
  locale:
    name: locale
    required: true
    in: path
    description: Used for detecting and setting the UI language.
    type: string
    enum: [ en, es, de, fr ] 
    # @TODO: Move locale (enum) to app or context config
  id:
    name: id
    required: true
    description: The id (as provided by the database) for an instance of a nodeType or a relationshipType.
    in: path
    type: integer
...
  title:
    name: title
    required: true
    in: formData
    description: The (main) title of the movie.
    type: string
  tagline:
    name: tagline
    required: true
    in: formData
    description: The tagline, slogan or subtitle which the movie has.
    type: string
  released:
    name: released
    required: true
    in: formData
    description: They year YYYY which the movie was released in.
    type: integer

### Persons ###
  name:
    name: name
    required: true
    in: formData
    description: a new persons full name
    type: string
  born:
    name: born
    required: true
    in: formData
    description: a new persons birth date
    type: integer

### Relationships ###
  source:
    name: source
    required: true
    in: formData
    description: A new relationships startpoint (person)
    type: string
  type:  
    name: type
    required: true
    in: formData
    description: A new relationships type
    type: string
  target:
    name: target
    required: true
    in: formData
    description: A new relationships endpoint (movie or person)
    type: string
  property:
    name: property
    required: false
    in: formData
    description: A new relationships property (optional and only for the types ACTED_IN and REVIEWED)
    type: string

## Complex Objects ##
definitions:
...
  MoviesListBody:
    required:
      - movies
    properties:
      movies:
        type: array
        items:
          $ref: "#/definitions/Movie"
  Movie:
    type: object
    required:
      # - tagline
      - title
      # - released
    properties:
      tagline:
        type: string
      title:
        type: string
      released:
        type: integer
      # rank:
      #   type: integer
      #   default: 5
      #   minimum: 1
      #   maximum: 10
  MovieBody:
    required:
      - movie
    properties:
      movie:
        $ref: "#/definitions/Movie"

### Persons ###
  PersonsListBody:
    required:
      - persons
    properties:
      persons:
        type: array
        items:
          $ref: "#/definitions/Person"
  Person:
    type: object
    required:
      - name
      # - born
    properties:
      name:
        type: string
      born:
        type: integer
      rank:
        type: integer
        default: 5
        minimum: 1
        maximum: 10
  PersonBody:
    required:
      - person
    properties:
      person:
        $ref: "#/definitions/Person"

### Responses ###
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponseWeb:
    # description: Error
    type: string